class changes. Main Application controller (name) is stored inside the Dispatcher. Added method "clear", which removes the last added object in the stack, and removes the object that is set to be dispatched.
invoke method changes.
Now accepts NULL for controller name. This is in case that a controller dispatch has been pretend and stored in the stack.
Before a controller is being called externally by another one not as a "main application controller", a check is being perform on a constant member.

pretend method changes.
Now optionaly can remove the Controller that is being tested from the controllers` stack. This is done automatically in case of failure.